All material-related modules moved to a materials package.
