--- solcore/material_data/__init__.py
@@ -1,1 +1,0 @@
-from .mobility import calculate_mobility
