--- solcore/material_system/__init__.py
@@ -1,2 +1,0 @@
-from .material_system import MaterialSystem
-from .create_new_material import create_new_material
