+++ solcore/materials/critical_point_picker.py
