+++ solcore/materials/refractiveindex_info_DB/dbmaterial.py
