--- solcore/parameter_system/__init__.py
@@ -1,1 +1,0 @@
-from .parameter_system import ParameterSystem
