--- solcore/units_system/__init__.py
@@ -1,1 +1,0 @@
-from .units_system import UnitsSystem
